% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_aggregate.R
\name{wid_aggregate}
\alias{wid_aggregate}
\title{Aggregate Dataset}
\usage{
wid_aggregate(data_dir, ..., levels = list(), cell_cutoff = 30,
  cores = parallel::detectCores() - 2)
}
\arguments{
\item{data_dir}{Path to the dataset, to be opened by \code{\link{wid_open}}.}

\item{...}{Arguments to pass to \code{\link{wid_subset}}.}

\item{levels}{A list containing new or redefined summaries to include.
The name of each entry determines the \code{variable} value, with named vectors
defining each \code{level} (e.g., \code{list(age = c(`Under 20` = "age < 20",
`Over 60` = "age > 60"))}).}

\item{cell_cutoff}{Minimal cell size to retain; cells with \code{count < cell_cutoff}
will be removed.}

\item{cores}{Number of CPU cores to use.}
}
\value{
An aggregated version of \code{db}.
}
\description{
Create aggregates from the dataset.
}
\examples{
\dontrun{

# aggregate a subset of the full dataset
data <- wid_open() |> dplyr::filter(country == "AGO")
aggregated_ago <- wid_aggregate(data)
}
}
