% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_convert_isic.R
\name{wid_convert_isic}
\alias{wid_convert_isic}
\title{Convert ISIC Codes}
\usage{
wid_convert_isic(codes, level = 4L, full_label = FALSE)
}
\arguments{
\item{codes}{A character vector of original ISIC codes, which can be at the
section, division, class, or 4-digit level. The latest unique classification
will be used unless the original code is prefixed with \code{40_},
\code{31_}, or \code{30_} for revisions \code{4}, \code{3.1}, or \code{3}.}

\item{level}{Level of dis-aggregation (determining number of categories):
\tabular{ll}{
  \code{1} \tab Section-level (default; 21 categories).\cr
  \code{2} \tab Lowest activity-level (6 categories).\cr
  \code{3} \tab 4 categories.\cr
  \code{4} \tab 3 categories.\cr
  \code{5} \tab Agricultural or not (2 categories).\cr
}}

\item{full_label}{Logical; if \code{TRUE}, will return full text labels
rather than short codes.}
}
\value{
A character vector containing short code or labels.
}
\description{
Convert International Standard Industrial Classification (ISIC) codes to
aggregate economic activity categories.
}
\examples{
(codes <- LETTERS[1:21])
wid_convert_isic(codes)
wid_convert_isic(codes, 1, full_label = TRUE)
wid_convert_isic(codes, 2, full_label = TRUE)
}
