% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_make_report.R
\name{wid_make_report}
\alias{wid_make_report}
\title{Report Variable Summaries}
\usage{
wid_make_report(reformat_dir, output = "vignettes/articles/variables.Rmd")
}
\arguments{
\item{reformat_dir}{Path to the reformatted dataset.}

\item{output}{Path to write the report to.}
}
\value{
Nothing; creates \code{output} file.
}
\description{
Generate a report of variable summaries.
}
\examples{
\dontrun{

  wid_make_report("../gender_growth_gap")
}
}
