% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_make_summaries.R
\name{wid_make_summaries}
\alias{wid_make_summaries}
\title{Summarize Dataset}
\usage{
wid_make_summaries(reformat_dir, manifest_dir = NULL, by_sex = FALSE,
  cores = parallel::detectCores() - 5)
}
\arguments{
\item{reformat_dir}{Directory containing the reformatted dataset.}

\item{manifest_dir}{Directory to write summaries to.}

\item{by_sex}{Logical; if \code{TRUE}, will further break down summaries by sex.}

\item{cores}{Number of CPU cores to use during processing.}
}
\value{
A list with entries for each variable:
\itemize{
  \item \code{type} Value type.
  \item \code{missing} Number of missing values across the entire dataset.
  \item \code{value_summary} Summary of the values, depending on type.
    For categorical variables, this will be observation counts for each level.
    For continuous variables, this will be the min, mean, standard deviation,
    median, and max. For identifier-like variables (character vectors with
    many unique values), this will be the alphabetically first and last value,
    and the number of unique values.
  \item \code{by_survey} Value summaries within each survey and year which
    have any observations.
}
}
\description{
Calculate summaries of the reformatted dataset.
}
\examples{
\dontrun{

  wid_make_summaries("../gender_growth_gap", "explore/public", by_sex = TRUE)
}
}
