% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_open.R
\name{wid_open}
\alias{wid_open}
\title{Open Dataset}
\usage{
wid_open(location = "work-in-data", ...,
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"))
}
\arguments{
\item{location}{Local or remote path to the dataset.}

\item{...}{Passes additional arguments to \code{\link[arrow]{s3_bucket}}.}

\item{key}{Your AWS access key; defaults to the \code{AWS_ACCESS_KEY_ID} environment variable.}

\item{secret}{Your AWS secret key; defaults to the
\code{AWS_SECRET_ACCESS_KEY} environment variable.}
}
\value{
The Work in Data \code{\link[arrow]{Dataset}}.
}
\description{
Open the Work in Data dataset.
}
\examples{
\dontrun{

  wid_open("/gpfs/gibbs/pi/dissc/work-in-data/gender_growth_gap")
}
}
