% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_schema.R
\name{wid_schema}
\alias{wid_schema}
\title{Work In Data Schema}
\usage{
wid_schema(original = FALSE)
}
\arguments{
\item{original}{If \code{TRUE}, returns the schema of the original
format, rather than that of the reformatted dataset.}
}
\value{
An Arrow \code{\link[arrow]{schema}}.
}
\description{
Make an Arrow schema to read the Work In Data files.
}
\examples{
wid_schema()
}
