% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_subset.R
\name{wid_subset}
\alias{wid_subset}
\title{Subset Dataset}
\usage{
wid_subset(db = NULL, ..., variables = NULL, surveys = NULL,
  min_year = -Inf, max_year = Inf, selection = NULL)
}
\arguments{
\item{db}{The dataset to subset; an \code{\link[arrow]{Dataset}}
as returned from \code{\link{wid_open}}.}

\item{...}{Conditions in the form of \code{variable operator value} (e.g., \code{age > 50}).}

\item{variables}{A character vector containing the variables you want to require.
That is, only observations with values for these variables will be included.}

\item{surveys}{A character vector of survey abbreviations to include.}

\item{min_year}{Earliest year to include.}

\item{max_year}{Latest year to include.}

\item{selection}{A character vector with country_year_survey IDs to be selected,
or a matrix-like object consisting of 3 columns containing country, year, and survey IDs
in that order.}
}
\value{
A filtered version of \code{db} if it is provided, or a list with entries
for remaining \code{variables}, \code{surveys}, and \code{years}.
}
\description{
Select a subset of the dataset.
}
\examples{
# see what's available in a given year
wid_subset(max_year = 1998)

# see which variables and years are in a particular survey
wid_subset(surveys = "CASEN")

# see which surveys and years have a particular variable
wid_subset(variables = "work_search")[-1]

# filter to a subset of data
db_dir <- "../../../gender_growth_gap"
if (dir.exists(db_dir)) {
  wid_open(db_dir) |> wid_subset(age > 200) |> dplyr::collect()
}
}
