% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_update_country_map.R, R/wid_update_isic.R,
%   R/wid_update_resources.R, R/wid_update_world_bank.R
\name{wid_update_country_map}
\alias{wid_update_country_map}
\alias{wid_update_isic}
\alias{wid_update_resources}
\alias{wid_update_world_bank}
\title{Update Resources}
\usage{
wid_update_country_map(out_dir = getOption("WID_RESOURCE_DIR"),
  overwrite = FALSE, simplify = FALSE)

wid_update_isic(out_dir = getOption("WID_RESOURCE_DIR"), overwrite = FALSE)

wid_update_resources(out_dir = getOption("WID_RESOURCE_DIR"),
  overwrite = FALSE)

wid_update_world_bank(out_dir = getOption("WID_RESOURCE_DIR"),
  overwrite = FALSE)
}
\arguments{
\item{out_dir}{Directory in which to save resources.}

\item{overwrite}{Logical; if \code{TRUE}, will re-download and replace existing resources.}

\item{simplify}{Logical; if \code{TRUE}, will simplify the shapes.}
}
\value{
Nothing; writes files to \code{out_dir}.
}
\description{
Download resources from external sources:
\itemize{
  \item \code{wid_update_resources}: Update all resources.
  \item \code{wid_update_world_bank}: Country-level data from the World Bank (using the \href{https://vincentarelbundock.github.io/WDI}{WDI package}).
  \item \code{wid_update_country_map}: GeoJSON map of countries from \href{https://github.com/nvkelso/natural-earth-vector}{Natural Earth Vector}.
  \item \code{wid_update_isic}: International Standard Industrial Classification (ISIC)
    code mappings from the \href{https://unstats.un.org/unsd/classifications/Econ/}{United Nations Statistics Division}.
}
}
\examples{
\dontrun{

  wid_update_resources()
}
}
