% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_variables.R
\name{wid_variables}
\alias{wid_variables}
\title{Work In Data Variables}
\usage{
wid_variables()
}
\value{
A list of variable info:
\itemize{
  \item \code{name}: Display name.
  \item \code{description}: Description of the variable.
  \item \code{type}: Arrow data type.
  \item \code{summary_type}: Type of the summary.
  \item \code{summary}: A summary, as created by \code{\link{wid_make_summaries}}.
}
}
\description{
Retrieve information about variables.
}
\examples{
names(wid_variables())
wid_variables()[["work"]][c(
  "name", "category", "description", "type", "conversion", "summary_type"
)]
}
