% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkInData-package.R
\docType{package}
\name{WorkInData-package}
\alias{WorkInData}
\alias{WorkInData-package}
\title{WorkInData: Work in Data}
\description{
Reformat and work with the Work in Data dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dissc-yale.github.io/WorkInData}
  \item \url{https://github.com/dissc-yale/WorkInData}
  \item Report bugs at \url{https://github.com/dissc-yale/WorkInData/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Other contributors:
\itemize{
  \item Yale Economic Growth Center [funder, copyright holder]
}

}
\keyword{internal}
