% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wid_convert_isic.R
\name{wid_convert_isic}
\alias{wid_convert_isic}
\title{Convert ISIC Codes}
\usage{
wid_convert_isic(codes, level = 4L, full_label = FALSE)
}
\arguments{
\item{codes}{A character vector of original ISIC codes, which can be at the
section, division, group, class, or 4-digit level. The latest unique classification
will be used unless the original code has a revision and/or code type prefix:
\code{revision_type_code}, where \code{revision} is one of
\code{40_}, \code{31_}, or \code{30_} for revisions \code{4}, \code{3.1}, or \code{3},
and \code{type} is one of \code{division}, \code{group}, or \code{class}.}

\item{level}{Level of dis-aggregation (determining number of categories):
\tabular{ll}{
  \code{1} \tab Section-level (21 categories).\cr
  \code{2} \tab Lowest activity-level (6 categories).\cr
  \code{3} \tab Split service sectors (4 categories).\cr
  \code{4} \tab Agriculture, Industry, or Services (3 categories).\cr
  \code{5} \tab Agricultural or not (2 categories).\cr
}}

\item{full_label}{Logical; if \code{TRUE}, will return full text labels
rather than short codes.}
}
\value{
A character vector containing short code or labels.
}
\description{
Convert International Standard Industrial Classification (ISIC) codes to
aggregate economic activity categories, based on the
\href{https://ilostat.ilo.org/methods/concepts-and-definitions/classification-economic-activities/}{International Labour Organization classification}.
}
\examples{
# revision 4 sections by default
(codes <- LETTERS[1:21])
wid_convert_isic(codes)
wid_convert_isic(codes, 1, full_label = TRUE)
wid_convert_isic(codes, 2, full_label = TRUE)

# disambiguating in the case of overlaps
wid_convert_isic("111", 4, full_label = TRUE)
wid_convert_isic("31_111", 4, full_label = TRUE)
wid_convert_isic("31_group_111", 4, full_label = TRUE)
}
